clang tidy fixes

+++ fuse_variables/test/test_acceleration_angular_2d_stamped.cpp
@@ -123,7 +123,7 @@ TEST(AccelerationAngular2DStamped, Optimization)
 
   // Build the problem.
   ceres::Problem problem;
-  problem.AddParameterBlock(acceleration.data(), acceleration.size());
+  problem.AddParameterBlock(acceleration.data(), static_cast<int>(acceleration.size()));
   std::vector<double*> parameter_blocks;
   parameter_blocks.push_back(acceleration.data());
   problem.AddResidualBlock(cost_function, nullptr, parameter_blocks);
