+++ fuse_variables/test/test_acceleration_angular_3d_stamped.cpp
@@ -127,7 +127,7 @@ TEST(AccelerationAngular3DStamped, Optimization)
 
   // Build the problem.
   ceres::Problem problem;
-  problem.AddParameterBlock(acceleration.data(), acceleration.size());
+  problem.AddParameterBlock(acceleration.data(), static_cast<int>(acceleration.size()));
   std::vector<double*> parameter_blocks;
   parameter_blocks.push_back(acceleration.data());
   problem.AddResidualBlock(cost_function, nullptr, parameter_blocks);
