+++ fuse_variables/test/test_velocity_angular_3d_stamped.cpp
@@ -127,7 +127,7 @@ TEST(VelocityAngular3DStamped, Optimization)
 
   // Build the problem.
   ceres::Problem problem;
-  problem.AddParameterBlock(velocity.data(), velocity.size());
+  problem.AddParameterBlock(velocity.data(), static_cast<int>(velocity.size()));
   std::vector<double*> parameter_blocks;
   parameter_blocks.push_back(velocity.data());
   problem.AddResidualBlock(cost_function, nullptr, parameter_blocks);
